new API

+++ database/migrations/2022_06_21_081324_create_pengumpulan_sampah.php
@@ -17,7 +17,7 @@ class CreatePengumpulanSampah extends Migration
             $table->increments('id');
             $table->integer('user_id')->unsigned();
             $table->double('jumlah');
-            $table->integer('point');
+            $table->double('point');
             $table->timestamps();
             $table->foreign('user_id')->references('id')->on('users');
         });
